LPS-53961 Rename test files for consistency. If we're testing a PermissionChecker name the test *PermissionCheckerTest
