+++ portal-impl/test/integration/com/liferay/portlet/messageboards/service/permission/MBCategoryPermissionCheckerTest.java
@@ -26,9 +26,7 @@ import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
 import com.liferay.portlet.messageboards.model.MBCategory;
 import com.liferay.portlet.messageboards.model.MBCategoryConstants;
-import com.liferay.portlet.messageboards.model.MBMessage;
 import com.liferay.portlet.messageboards.service.MBCategoryLocalServiceUtil;
-import com.liferay.portlet.messageboards.service.MBMessageLocalServiceUtil;
 
 import org.junit.Assert;
 import org.junit.ClassRule;
@@ -39,7 +37,7 @@ import org.junit.Test;
  * @author Eric Chin
  * @author Shinn Lok
  */
-public class MBMessagePermissionTest extends BasePermissionTestCase {
+public class MBCategoryPermissionCheckerTest extends BasePermissionTestCase {
 
 	@ClassRule
 	@Rule
@@ -50,20 +48,20 @@ public class MBMessagePermissionTest extends BasePermissionTestCase {
 	@Test
 	public void testContains() throws Exception {
 		Assert.assertTrue(
-			MBMessagePermission.contains(
-				permissionChecker, _message, ActionKeys.VIEW));
+			MBCategoryPermission.contains(
+				permissionChecker, _category, ActionKeys.VIEW));
 		Assert.assertTrue(
-			MBMessagePermission.contains(
-				permissionChecker, _submessage, ActionKeys.VIEW));
+			MBCategoryPermission.contains(
+				permissionChecker, _subcategory, ActionKeys.VIEW));
 
 		removePortletModelViewPermission();
 
 		Assert.assertFalse(
-			MBMessagePermission.contains(
-				permissionChecker, _message, ActionKeys.VIEW));
+			MBCategoryPermission.contains(
+				permissionChecker, _category, ActionKeys.VIEW));
 		Assert.assertFalse(
-			MBMessagePermission.contains(
-				permissionChecker, _submessage, ActionKeys.VIEW));
+			MBCategoryPermission.contains(
+				permissionChecker, _subcategory, ActionKeys.VIEW));
 	}
 
 	@Override
@@ -72,22 +70,14 @@ public class MBMessagePermissionTest extends BasePermissionTestCase {
 			ServiceContextTestUtil.getServiceContext(
 				group.getGroupId(), TestPropsValues.getUserId());
 
-		_message = MBMessageLocalServiceUtil.addMessage(
-			TestPropsValues.getUserId(), RandomTestUtil.randomString(),
-			group.getGroupId(), MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID,
-			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
-			serviceContext);
-
-		MBCategory category = MBCategoryLocalServiceUtil.addCategory(
+		_category = MBCategoryLocalServiceUtil.addCategory(
 			TestPropsValues.getUserId(),
 			MBCategoryConstants.DEFAULT_PARENT_CATEGORY_ID,
 			RandomTestUtil.randomString(), StringPool.BLANK, serviceContext);
 
-		_submessage = MBMessageLocalServiceUtil.addMessage(
-			TestPropsValues.getUserId(), RandomTestUtil.randomString(),
-			group.getGroupId(), category.getCategoryId(),
-			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
-			serviceContext);
+		_subcategory = MBCategoryLocalServiceUtil.addCategory(
+			TestPropsValues.getUserId(), _category.getCategoryId(),
+			RandomTestUtil.randomString(), StringPool.BLANK, serviceContext);
 	}
 
 	@Override
@@ -95,7 +85,7 @@ public class MBMessagePermissionTest extends BasePermissionTestCase {
 		return MBPermission.RESOURCE_NAME;
 	}
 
-	private MBMessage _message;
-	private MBMessage _submessage;
+	private MBCategory _category;
+	private MBCategory _subcategory;
 
 }
