+++ portal-impl/test/integration/com/liferay/portlet/shopping/service/permission/ShoppingCategoryPermissionCheckerTest.java
@@ -20,7 +20,6 @@ import com.liferay.portal.service.permission.test.BasePermissionTestCase;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.MainServletTestRule;
 import com.liferay.portlet.shopping.model.ShoppingCategory;
-import com.liferay.portlet.shopping.model.ShoppingItem;
 import com.liferay.portlet.shopping.util.test.ShoppingTestUtil;
 
 import org.junit.Assert;
@@ -32,7 +31,8 @@ import org.junit.Test;
  * @author Eric Chin
  * @author Shinn Lok
  */
-public class ShoppingItemPermissionTest extends BasePermissionTestCase {
+public class ShoppingCategoryPermissionCheckerTest
+	extends BasePermissionTestCase {
 
 	@ClassRule
 	@Rule
@@ -43,31 +43,28 @@ public class ShoppingItemPermissionTest extends BasePermissionTestCase {
 	@Test
 	public void testContains() throws Exception {
 		Assert.assertTrue(
-			ShoppingItemPermission.contains(
-				permissionChecker, _item, ActionKeys.VIEW));
+			ShoppingCategoryPermission.contains(
+				permissionChecker, _category, ActionKeys.VIEW));
 		Assert.assertTrue(
-			ShoppingItemPermission.contains(
-				permissionChecker, _subitem, ActionKeys.VIEW));
+			ShoppingCategoryPermission.contains(
+				permissionChecker, _subcategory, ActionKeys.VIEW));
 
 		removePortletModelViewPermission();
 
 		Assert.assertFalse(
-			ShoppingItemPermission.contains(
-				permissionChecker, _item, ActionKeys.VIEW));
+			ShoppingCategoryPermission.contains(
+				permissionChecker, _category, ActionKeys.VIEW));
 		Assert.assertFalse(
-			ShoppingItemPermission.contains(
-				permissionChecker, _subitem, ActionKeys.VIEW));
+			ShoppingCategoryPermission.contains(
+				permissionChecker, _subcategory, ActionKeys.VIEW));
 	}
 
 	@Override
 	protected void doSetUp() throws Exception {
-		_item = ShoppingTestUtil.addItem(group.getGroupId());
+		_category = ShoppingTestUtil.addCategory(group.getGroupId());
 
-		ShoppingCategory category = ShoppingTestUtil.addCategory(
-			group.getGroupId());
-
-		_subitem = ShoppingTestUtil.addItem(
-			group.getGroupId(), category.getCategoryId());
+		_subcategory = ShoppingTestUtil.addCategory(
+			group.getGroupId(), _category.getCategoryId());
 	}
 
 	@Override
@@ -75,7 +72,7 @@ public class ShoppingItemPermissionTest extends BasePermissionTestCase {
 		return ShoppingPermission.RESOURCE_NAME;
 	}
 
-	private ShoppingItem _item;
-	private ShoppingItem _subitem;
+	private ShoppingCategory _category;
+	private ShoppingCategory _subcategory;
 
 }
