+++ portal-impl/test/integration/com/liferay/portlet/shopping/service/permission/ShoppingItemPermissionCheckerTest.java
@@ -12,72 +12,70 @@
  * details.
  */
 
-package com.liferay.journal.service.permission.test;
+package com.liferay.portlet.shopping.service.permission;
 
-import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.journal.test.util.JournalTestUtil;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.security.permission.ActionKeys;
 import com.liferay.portal.service.permission.test.BasePermissionTestCase;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portlet.journal.model.JournalFolder;
-import com.liferay.portlet.journal.service.permission.JournalFolderPermission;
-import com.liferay.portlet.journal.service.permission.JournalPermission;
+import com.liferay.portal.test.rule.MainServletTestRule;
+import com.liferay.portlet.shopping.model.ShoppingCategory;
+import com.liferay.portlet.shopping.model.ShoppingItem;
+import com.liferay.portlet.shopping.util.test.ShoppingTestUtil;
 
 import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 /**
  * @author Eric Chin
  * @author Shinn Lok
  */
-@RunWith(Arquillian.class)
-public class JournalFolderPermissionTest extends BasePermissionTestCase {
+public class ShoppingItemPermissionCheckerTest extends BasePermissionTestCase {
 
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule =
-		new LiferayIntegrationTestRule();
+		new AggregateTestRule(
+			new LiferayIntegrationTestRule(), MainServletTestRule.INSTANCE);
 
 	@Test
 	public void testContains() throws Exception {
 		Assert.assertTrue(
-			JournalFolderPermission.contains(
-				permissionChecker, _folder, ActionKeys.VIEW));
+			ShoppingItemPermission.contains(
+				permissionChecker, _item, ActionKeys.VIEW));
 		Assert.assertTrue(
-			JournalFolderPermission.contains(
-				permissionChecker, _subfolder, ActionKeys.VIEW));
+			ShoppingItemPermission.contains(
+				permissionChecker, _subitem, ActionKeys.VIEW));
 
 		removePortletModelViewPermission();
 
 		Assert.assertFalse(
-			JournalFolderPermission.contains(
-				permissionChecker, _folder, ActionKeys.VIEW));
+			ShoppingItemPermission.contains(
+				permissionChecker, _item, ActionKeys.VIEW));
 		Assert.assertFalse(
-			JournalFolderPermission.contains(
-				permissionChecker, _subfolder, ActionKeys.VIEW));
+			ShoppingItemPermission.contains(
+				permissionChecker, _subitem, ActionKeys.VIEW));
 	}
 
 	@Override
 	protected void doSetUp() throws Exception {
-		_folder = JournalTestUtil.addFolder(
-			group.getGroupId(), RandomTestUtil.randomString());
+		_item = ShoppingTestUtil.addItem(group.getGroupId());
 
-		_subfolder = JournalTestUtil.addFolder(
-			group.getGroupId(), _folder.getFolderId(),
-			RandomTestUtil.randomString());
+		ShoppingCategory category = ShoppingTestUtil.addCategory(
+			group.getGroupId());
+
+		_subitem = ShoppingTestUtil.addItem(
+			group.getGroupId(), category.getCategoryId());
 	}
 
 	@Override
 	protected String getResourceName() {
-		return JournalPermission.RESOURCE_NAME;
+		return ShoppingPermission.RESOURCE_NAME;
 	}
 
-	private JournalFolder _folder;
-	private JournalFolder _subfolder;
+	private ShoppingItem _item;
+	private ShoppingItem _subitem;
 
 }
